シミュレーションデータの表示・非表示
